+++ docs/source/whats-new.rst
@@ -46,7 +46,7 @@ Deprecations
 Bug fixes
 ~~~~~~~~~
 
-- Renaming a node also update the key of its parent's children
+- Renaming a node also updates the key of its parent's children
   (:issue:`309`, :pull:`310`)
   By `Etienne Schalk <https://github.com/etienneschalk>`_.
 - Keep attributes on nodes containing no data in :py:func:`map_over_subtree`. (:issue:`278`, :pull:`279`)